[REFACTOR] moving mew models to the mew sub-package to improve consistency
