+++ src/mewpy/io/engines/cobra_model.py
@@ -3,12 +3,12 @@ from typing import Union, TYPE_CHECKING
 
 from mewpy.io.dto import VariableRecord, DataTransferObject, CompartmentRecord, FunctionTerm
 from mewpy.mew.algebra import Expression
-from mewpy.model import MetabolicModel
+from mewpy.mew.models import MetabolicModel
 from .engine import Engine
 from .engines_utils import build_symbolic, expression_warning, cobra_warning
 
 if TYPE_CHECKING:
-    from mewpy.model import RegulatoryModel, Model, MetabolicModel
+    from mewpy.mew.models import RegulatoryModel, Model, MetabolicModel
 
 
 class CobraModel(Engine):
