+++ src/mewpy/io/engines/json.py
@@ -2,13 +2,13 @@ import json
 import os
 from typing import Union, TYPE_CHECKING
 
-from mewpy.model import Model
+from mewpy.mew.models import Model
 from mewpy.io.dto import DataTransferObject
 
 from .engine import Engine
 
 if TYPE_CHECKING:
-    from mewpy.model import RegulatoryModel, Model, MetabolicModel
+    from mewpy.mew.models import RegulatoryModel, Model, MetabolicModel
 
 
 class JSON(Engine):
