+++ src/mewpy/mew/analysis/pfba.py
@@ -2,7 +2,7 @@ from typing import Union, Dict
 
 from mewpy.mew.analysis import FBA
 from mewpy.mew.lp import ConstraintContainer, VariableContainer
-from mewpy.model import Model, MetabolicModel, RegulatoryModel
+from mewpy.mew.models import Model, MetabolicModel, RegulatoryModel
 from mewpy.solvers.solution import Solution, Status
 from mewpy.solvers.solver import VarType, Solver
