+++ src/mewpy/mew/analysis/prom.py
@@ -8,7 +8,7 @@ from mewpy.util.constants import ModelConstants
 
 if TYPE_CHECKING:
     from mewpy.mew.variables import Regulator, Gene, Target
-    from mewpy.model import Model, MetabolicModel, RegulatoryModel
+    from mewpy.mew.models import Model, MetabolicModel, RegulatoryModel
 
 
 def _run_and_decode_solver(lp,
