+++ src/mewpy/mew/analysis/rfba.py
@@ -4,7 +4,7 @@ from warnings import warn
 from mewpy.mew.analysis import FBA
 from mewpy.solvers import Solution
 from mewpy.solvers.solver import Solver
-from mewpy.model import Model, MetabolicModel, RegulatoryModel
+from mewpy.mew.models import Model, MetabolicModel, RegulatoryModel
 from mewpy.util.constants import ModelConstants
 from mewpy.mew.solution import ModelSolution, DynamicSolution
