+++ src/mewpy/mew/models/__init__.py
@@ -1,0 +1,3 @@
+from .model import Model, build_model
+from .metabolic import MetabolicModel
+from .regulatory import RegulatoryModel
