+++ src/mewpy/mew/models/metabolic.py
@@ -1,9 +1,9 @@
 from typing import TYPE_CHECKING, Any, Union, Generator, Dict, List, Tuple, Set
 
-from mewpy.model.model import Model
+from .model import Model
 from mewpy.util.history import recorder
-from mewpy.util.serialization import serialize
-from mewpy.util.utilities import iterable, generator
+from mewpy.mew.models.serialization import serialize
+from mewpy.util.utilities import generator
 
 if TYPE_CHECKING:
     from mewpy.mew.algebra import Expression
