+++ src/mewpy/mew/models/regulatory.py
@@ -1,8 +1,8 @@
 from typing import Any, TYPE_CHECKING, Union, Generator, Dict, List, Tuple, Set
 
-from mewpy.model.model import Model
+from .model import Model
 from mewpy.util.history import recorder
-from mewpy.util.serialization import serialize
+from mewpy.mew.models.serialization import serialize
 from mewpy.util.utilities import generator
 
 if TYPE_CHECKING:
