+++ src/mewpy/mew/solution/model_solution.py
@@ -9,7 +9,7 @@ if TYPE_CHECKING:
     from mewpy.solvers import Solution
     from mewpy.mew.variables import Variable
     from mewpy.mew.lp import LinearProblem
-    from mewpy.model import Model, MetabolicModel, RegulatoryModel
+    from mewpy.mew.models import Model, MetabolicModel, RegulatoryModel
 
 
 class ModelSolution:
