+++ src/mewpy/mew/variables/gene.py
@@ -5,7 +5,7 @@ from mewpy.util.constants import ModelConstants
 from mewpy.util.history import recorder
 
 from mewpy.util.utilities import generator
-from mewpy.util.serialization import serialize
+from mewpy.mew.models.serialization import serialize
 from .variable import Variable
 from .variables_utils import coefficients_setter
