+++ src/mewpy/mew/variables/interaction.py
@@ -4,7 +4,7 @@ import pandas as pd
 
 from mewpy.mew.algebra import Expression, parse_expression
 from mewpy.util.utilities import generator
-from mewpy.util.serialization import serialize
+from mewpy.mew.models.serialization import serialize
 from mewpy.util.history import recorder
 from mewpy.io.engines.engines_utils import expression_warning
 from .variable import Variable, variables_from_symbolic
@@ -14,7 +14,7 @@ from .variable import Variable, variables_from_symbolic
 if TYPE_CHECKING:
     from .regulator import Regulator
     from .target import Target
-    from mewpy.model import Model, MetabolicModel, RegulatoryModel
+    from mewpy.mew.models import Model, MetabolicModel, RegulatoryModel
 
 
 class Interaction(Variable, variable_type='interaction', register=True, constructor=True, checker=True):
