+++ src/mewpy/model/__init__.py
@@ -1,6 +1,3 @@
-from .model import Model, build_model
-from .metabolic import MetabolicModel
-from .regulatory import RegulatoryModel
 from .gecko import GeckoModel
 from .smoment import SMomentModel
 from .com import CommunityModel
