+++ src/mewpy/omics/integration/coregflux.py
@@ -5,7 +5,7 @@ from mewpy.util.constants import ModelConstants
 from mewpy.omics import ExpressionSet
 
 if TYPE_CHECKING:
-    from mewpy.model import Model, MetabolicModel, RegulatoryModel
+    from mewpy.mew.models import Model, MetabolicModel, RegulatoryModel
 
 
 def CoRegFlux(model: Union['Model', 'MetabolicModel', 'RegulatoryModel'],
