+++ src/mewpy/simulation/__init__.py
@@ -49,11 +49,11 @@ map_model_simulator = {
     'geckopy.gecko.GeckoModel': ('mewpy.simulation.cobra', 'GeckoSimulation'),
     'mewpy.model.gecko.GeckoModel': ('mewpy.simulation.reframed', 'GeckoSimulation'),
     'mewpy.model.smoment.SMomentModel': ('mewpy.simulation.reframed', 'GeckoSimulation'),
-    'mewpy.model.model.Model': ('mewpy.simulation.mew', 'Simulation'),
-    'mewpy.model.metabolic.MetabolicModel': ('mewpy.simulation.mew', 'Simulation'),
-    'mewpy.model.regulatory.RegulatoryModel': ('mewpy.simulation.mew', 'Simulation'),
-    'mewpy.model.model.MetabolicRegulatoryModel': ('mewpy.simulation.mew', 'Simulation'),
-    'mewpy.model.model.RegulatoryMetabolicModel': ('mewpy.simulation.mew', 'Simulation'),
+    'mewpy.mew.models.model.Model': ('mewpy.simulation.mew', 'Simulation'),
+    'mewpy.mew.models.metabolic.MetabolicModel': ('mewpy.simulation.mew', 'Simulation'),
+    'mewpy.mew.models.regulatory.RegulatoryModel': ('mewpy.simulation.mew', 'Simulation'),
+    'mewpy.mew.models.model.MetabolicRegulatoryModel': ('mewpy.simulation.mew', 'Simulation'),
+    'mewpy.mew.models.model.RegulatoryMetabolicModel': ('mewpy.simulation.mew', 'Simulation'),
 }
 
 
@@ -132,13 +132,13 @@ def get_container(model):
 
     """
 
-    from mewpy.model import Model, RegulatoryModel, MetabolicModel
+    from mewpy.mew.models import Model, RegulatoryModel, MetabolicModel
 
     if isinstance(model, (Model, MetabolicModel, RegulatoryModel)):
 
-        from .mew import MewModelContainer
+        from .mew import MewModel
 
-        return MewModelContainer(model)
+        return MewModel(model)
 
     try:
         from reframed.core.cbmodel import CBModel
